requirements.txt
transformers>=4.30.0
torch>=2.0.0
gradio>=3.36.0
ffmpeg-python>=0.2.0
sentence-transformers>=2.2.2
Pillow>=9.5.0
python-dotenv>=1.0.0
openai>=1.0.0  # Add this line
scikit-image>=0.19.0
opencv-python>=4.5.0
numpy>=1.21.0